+++ include/continental/landscapeevolutionmodel/domain/StreamDefinitionConfig.h
@@ -23,11 +23,11 @@ class CONTINENTALLANDSCAPEEVOLUTIONMODELPLUGIN_EXPORT_DECL StreamDefinitionConfi
 public:
     /// Função de retorno do valor da área mínima.
     /// @return Retorna o valor da área mínima.
-    float getThresoldValue() const;
+    double getThresoldValue() const;
 
     /// Função de definição do valor da área mínima.
     /// @param thresoldValue Retorna o valor da área mínima.
-    void setThresoldValue(float thresoldValue);
+    void setThresoldValue(double thresoldValue);
 
     /// Função de retorno do enum do tipo de algoritmo.
     /// @return Retorna o enum do tipo de algoritmo.
@@ -43,7 +43,7 @@ public:
 
 private:
     /// Membros.
-    float m_thresoldValue = 0.0f; //Acho que o usuário deve poderalterar esse valor
+    double m_thresoldValue = 0.0f; //Acho que o usuário deve poderalterar esse valor
     StreamDefinitionThresholdType m_thresoldType;
 };
